Tenant moves to internal API
