+++ robozonky-app/src/main/java/com/github/robozonky/app/daemon/AmountsAtRiskSummarizer.java
@@ -22,7 +22,7 @@ import java.util.Map;
 
 import com.github.robozonky.api.remote.entities.sanitized.Investment;
 import com.github.robozonky.api.remote.enums.Rating;
-import com.github.robozonky.app.authentication.Tenant;
+import com.github.robozonky.common.Tenant;
 import com.github.robozonky.common.remote.Zonky;
 import com.github.robozonky.internal.util.BigDecimalCalculator;
 import org.slf4j.Logger;
