+++ robozonky-app/src/main/java/com/github/robozonky/app/daemon/BlockedAmountProcessor.java
@@ -30,7 +30,7 @@ import javax.ws.rs.NotFoundException;
 import com.github.robozonky.api.remote.entities.BlockedAmount;
 import com.github.robozonky.api.remote.entities.sanitized.Loan;
 import com.github.robozonky.api.remote.enums.Rating;
-import com.github.robozonky.app.authentication.Tenant;
+import com.github.robozonky.common.Tenant;
 import com.github.robozonky.common.remote.Zonky;
 import com.github.robozonky.internal.util.LazyInitialized;
 import org.slf4j.Logger;
