+++ robozonky-app/src/main/java/com/github/robozonky/app/daemon/DelinquencyCategory.java
@@ -36,9 +36,9 @@ import com.github.robozonky.api.notifications.LoanNowDelinquentEvent;
 import com.github.robozonky.api.remote.entities.sanitized.Development;
 import com.github.robozonky.api.remote.entities.sanitized.Investment;
 import com.github.robozonky.api.remote.entities.sanitized.Loan;
-import com.github.robozonky.app.authentication.Tenant;
 import com.github.robozonky.app.events.EventFactory;
 import com.github.robozonky.app.events.LazyEvent;
+import com.github.robozonky.common.Tenant;
 import com.github.robozonky.common.state.InstanceState;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
