+++ robozonky-app/src/main/java/com/github/robozonky/app/daemon/DelinquencyNotificationPayload.java
@@ -26,8 +26,8 @@ import java.util.stream.Stream;
 import com.github.robozonky.api.remote.entities.sanitized.Investment;
 import com.github.robozonky.api.remote.entities.sanitized.Loan;
 import com.github.robozonky.api.remote.enums.PaymentStatus;
-import com.github.robozonky.app.authentication.Tenant;
 import com.github.robozonky.app.authentication.TenantBuilder;
+import com.github.robozonky.common.Tenant;
 import com.github.robozonky.common.jobs.Payload;
 import com.github.robozonky.common.remote.Zonky;
 import com.github.robozonky.common.secrets.SecretProvider;
