+++ robozonky-app/src/main/java/com/github/robozonky/app/daemon/operations/Purchasing.java
@@ -26,9 +26,9 @@ import com.github.robozonky.api.remote.entities.Participation;
 import com.github.robozonky.api.remote.entities.sanitized.Investment;
 import com.github.robozonky.api.strategies.ParticipationDescriptor;
 import com.github.robozonky.api.strategies.PurchaseStrategy;
-import com.github.robozonky.app.authentication.Tenant;
 import com.github.robozonky.app.daemon.LoanCache;
 import com.github.robozonky.app.daemon.Portfolio;
+import com.github.robozonky.common.Tenant;
 import com.github.robozonky.util.NumberUtil;
 
 public class Purchasing extends StrategyExecutor<Participation, PurchaseStrategy> {
