+++ robozonky-app/src/main/java/com/github/robozonky/app/daemon/operations/Selling.java
@@ -29,12 +29,12 @@ import com.github.robozonky.api.strategies.InvestmentDescriptor;
 import com.github.robozonky.api.strategies.PortfolioOverview;
 import com.github.robozonky.api.strategies.RecommendedInvestment;
 import com.github.robozonky.api.strategies.SellStrategy;
-import com.github.robozonky.app.authentication.Tenant;
 import com.github.robozonky.app.daemon.LoanCache;
 import com.github.robozonky.app.daemon.Portfolio;
 import com.github.robozonky.app.daemon.PortfolioDependant;
 import com.github.robozonky.app.daemon.TransactionalPortfolio;
 import com.github.robozonky.app.events.EventFactory;
+import com.github.robozonky.common.Tenant;
 import com.github.robozonky.common.remote.Select;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
