+++ robozonky-app/src/main/java/com/github/robozonky/app/daemon/transactions/LoanRepaidProcessor.java
@@ -22,9 +22,9 @@ import com.github.robozonky.api.remote.entities.sanitized.Loan;
 import com.github.robozonky.api.remote.enums.PaymentStatus;
 import com.github.robozonky.api.remote.enums.TransactionCategory;
 import com.github.robozonky.api.remote.enums.TransactionOrientation;
-import com.github.robozonky.app.authentication.Tenant;
 import com.github.robozonky.app.daemon.LoanCache;
 import com.github.robozonky.app.daemon.TransactionalPortfolio;
+import com.github.robozonky.common.Tenant;
 
 import static com.github.robozonky.app.events.EventFactory.loanRepaid;
 import static com.github.robozonky.app.events.EventFactory.loanRepaidLazy;
