+++ robozonky-app/src/main/java/com/github/robozonky/app/daemon/transactions/SoldParticipationCache.java
@@ -28,7 +28,7 @@ import java.util.stream.Collectors;
 
 import com.github.robozonky.api.SessionInfo;
 import com.github.robozonky.api.remote.entities.sanitized.Investment;
-import com.github.robozonky.app.authentication.Tenant;
+import com.github.robozonky.common.Tenant;
 import com.github.robozonky.common.remote.Select;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
