+++ robozonky-app/src/test/java/com/github/robozonky/app/daemon/BlockedAmountProcessorTest.java
@@ -25,7 +25,7 @@ import com.github.robozonky.api.remote.entities.BlockedAmount;
 import com.github.robozonky.api.remote.entities.sanitized.Loan;
 import com.github.robozonky.api.remote.enums.Rating;
 import com.github.robozonky.app.AbstractZonkyLeveragingTest;
-import com.github.robozonky.app.authentication.Tenant;
+import com.github.robozonky.common.Tenant;
 import com.github.robozonky.common.remote.Zonky;
 import com.github.robozonky.internal.util.Maps;
 import org.junit.jupiter.api.Test;
