+++ robozonky-app/src/test/java/com/github/robozonky/app/daemon/PurchasingDaemonTest.java
@@ -22,7 +22,7 @@ import java.util.function.Supplier;
 
 import com.github.robozonky.api.strategies.PurchaseStrategy;
 import com.github.robozonky.app.AbstractZonkyLeveragingTest;
-import com.github.robozonky.app.authentication.Tenant;
+import com.github.robozonky.common.Tenant;
 import com.github.robozonky.common.remote.Zonky;
 import org.junit.jupiter.api.Test;
