+++ robozonky-app/src/test/java/com/github/robozonky/app/daemon/operations/SessionStateTest.java
@@ -21,7 +21,7 @@ import java.util.UUID;
 
 import com.github.robozonky.api.remote.entities.sanitized.Loan;
 import com.github.robozonky.app.AbstractZonkyLeveragingTest;
-import com.github.robozonky.app.authentication.Tenant;
+import com.github.robozonky.common.Tenant;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
